Estrutura sequencial - exercicio - 04 soma

+++ 01 estrutura sequencial/j04Soma.java
@@ -9,13 +9,13 @@ public class j04Soma {
   public static void main(String[] args) {
     
     Scanner input = new Scanner(System.in);
-    double x, y, soma;
+    int x, y, soma;
     
     System.out.print("Digite o valor de X: ");
-    x = input.nextDouble();
+    x = input.nextInt();
     
     System.out.print("Digite o valor de Y: ");
-    y = input.nextDouble();
+    y = input.nextInt();
 
     soma = x + y;
